res made the following changes:   - Changed the iCamera interface so that getting the projection     matrix doesn't require the screen dimensions any more.   - iProjectionCamera now takes FOV, perspective center etc. in     normalized screen space, not pixels. (Note: old iCamera still     takes pixel coords!)   - Added visible volume determination for custom matrix camera.   - Changed how RenderViewClipper works: it now takes the visible     volume from the camera instead of the frustum computed by the     render view.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/feature/camimproved@28945 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csgeom/projections.h
@@ -55,7 +55,7 @@ namespace CS
       /**
        * Create a perspective projection as used by CS.
        */
-      static Matrix4 CSPerspective (int viewWidth, int viewHeight, 
+      static Matrix4 CSPerspective (float viewWidth, float viewHeight, 
 				    float shiftX, float shiftY,
 				    float invAspect)
       {
